Feat: 이모지 추가시 예외 경고 Toast 추가
