+++ src/components/layout/subHeader/SubHeader.jsx
@@ -1,19 +1,31 @@
+import { useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import { useGetData } from '../../../hooks/useGetData';
 import { BASE_URL_RECIPIENT } from '../../../constants/url';
 import ProfileList from './ProfileList';
 import EmojiBox from './EmojiBox';
-import './subHeader.scss';
 import ShareBox from '../../share/ShareBox';
+import Toast from '../../toast/Toast';
+import './subHeader.scss';
 
 export default function SubHeader() {
   const { id } = useParams();
+  const [isShowToast, setIsShowToast] = useState(false);
+  const [messageText, setMessageText] = useState('');
+
   const { data: getAllapi, isLoading: allLoading } = useGetData(
     `${BASE_URL_RECIPIENT}${id}/`
   );
   const { data: getEmojiApi, isLoading: emojiLoading } = useGetData(
     `${BASE_URL_RECIPIENT}${id}/reactions/`
   );
+
+  // url 공유하기
+  const handleToast = (message) => {
+    setMessageText(message);
+    setIsShowToast(true);
+    setTimeout(() => setIsShowToast(false), 1500);
+  };
 
   const Loading = allLoading === false && emojiLoading === false;
 
@@ -30,11 +42,19 @@ export default function SubHeader() {
             </div>
             <div className="d__flex--center subHeader__right">
               {getAllapi && <ProfileList coment={getAllapi.recentMessages} />}
-              <EmojiBox emojis={getEmojiApi?.results} rollingPageId={id} />
-              <ShareBox name={getAllapi?.name || ''} />
+              <EmojiBox
+                emojis={getEmojiApi?.results}
+                rollingPageId={id}
+                onToastMessage={handleToast}
+              />
+              <ShareBox
+                name={getAllapi?.name || ''}
+                onToastMessage={handleToast}
+              />
             </div>
           </>
         )}
+        {isShowToast && <Toast message={messageText} />}
       </div>
     </section>
   );
